Workspace/NextGenATM/NextGenATM.Services/requirements.txt
boto3==1.26.89
cairosvg==2.6.0
face_recognition==1.3.0
Flask==2.2.3
flask_cors==3.0.10
flask_restful==0.3.9
Pillow==9.4.0
pymongo==4.3.3
pyqrcode==1.2.1
